requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
pymysql
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart
cos-python-sdk-v5
